group taxon/assembly attributes in a single table Fixes #10
