+++ src/client/views/components/Page.jsx
@@ -1,8 +1,10 @@
 import Grid from "@material-ui/core/Grid";
 import React from "react";
 import SearchBox from "./SearchBox";
+import classnames from "classnames";
 import { compose } from "recompose";
 import { makeStyles } from "@material-ui/core/styles";
+import styles from "./Styles.scss";
 
 const useStyles = makeStyles((theme) => ({
   container: {
@@ -13,7 +15,16 @@ const useStyles = makeStyles((theme) => ({
   item: { minWidth: "900px", maxWidth: "80%", align: "center" },
 }));
 
-const Page = ({ searchBox, panels, text, pageRef }) => {
+const Page = ({
+  searchBox,
+  panels,
+  text,
+  pageRef,
+  recordId,
+  fieldId,
+  resultCount,
+  result,
+}) => {
   const classes = useStyles();
   let items = [];
   if (panels && panels.length > 0) {
@@ -31,6 +42,16 @@ const Page = ({ searchBox, panels, text, pageRef }) => {
       );
     });
   }
+  let title;
+  if (recordId && result) {
+    title = `${result} record ${recordId}`;
+  } else if (fieldId) {
+    title = `${fieldId} summary`;
+  } else if (resultCount >= 0) {
+    title = `${resultCount} ${resultCount == 1 ? "hit" : "hits"}`;
+  } else if (resultCount < 0) {
+    title = `updating search results…`;
+  }
   return (
     <Grid
       container
@@ -44,6 +65,15 @@ const Page = ({ searchBox, panels, text, pageRef }) => {
       {searchBox && (
         <Grid item className={classes.item} style={{ marginTop: "2em" }}>
           <SearchBox />
+        </Grid>
+      )}
+      {title && (
+        <Grid
+          item
+          className={classnames(styles.pageTitle, classes.item)}
+          style={{ marginBottom: "0.5em" }}
+        >
+          {title}
         </Grid>
       )}
       {items}
