+++ src/client/views/components/Styles.scss
@@ -151,6 +151,19 @@ body,
       }
     }
 
+    .pageTitle {
+      color: themed("primary");
+      background-color: themed("secondary");
+      border-bottom: themed("primary") 0.1em solid;
+      font-family: themed("titleFont");
+      font-size: 1.25em;
+      text-align: left;
+      padding: 0.25em;
+      &:first-letter {
+        text-transform: capitalize;
+      }
+    }
+
     .searchBox {
       @include placeholderColor(
         mix(themed("primary"), themed("secondary"), 65%)
@@ -336,6 +349,10 @@ body,
 
       .header {
         cursor: pointer;
+        margin: -0.25em -0.25em 0.25em -0.25em;
+        padding: 0.25em;
+        color: themed("secondary");
+        background-color: themed("primary");
 
         .identifier {
           float: right;
